requirements.txt
# Atsu Downloader - Dependencies
# Install with: pip install -r requirements.txt

typer>=0.9.0
rich>=13.0.0
requests>=2.31.0
Pillow>=10.0.0
PyQt6>=6.5.0
